admin优化

+++ server/src/controller/user.js
@@ -121,68 +121,6 @@ class userController{
 		} else {
 			handleError({ ctx, msg: '获取列表数据失败' })
 		}
-
-		// let currentPage = parseInt(ctx.query.currentPage)
-		// let pageSize = parseInt(ctx.query.pageSize)
-		// let queryName = ctx.query.queryName
-		// let queryRole = parseInt(ctx.query.queryRole)
-		// console.log(ctx.query)
-		// // console.log(ctx.query.queryName)
-		// if(currentPage <= 0) currentPage = 1
-		// let result=null,total=0
-		// // 组合搜索
-    // const querys = {}
-    // // 名字查询
-    // if(queryName){
-    // 	querys['$or'] = [ 
-	  //     { 'username': {$regex: queryName} } // 使用正则模糊搜索
-	  //   ]
-    // }
-    // // 权限查询
-    // if (queryRole) { querys.role = queryRole }
-		// // 如何将密码去除在返回到前端？
-		// if(currentPage && pageSize){
-		// 	result = await User
-		// 		// .find({'username': new RegExp(ctx.query.queryName)}) // 模糊搜索
-		// 		// .find({'username': {$regex: ctx.query.queryName}}) // 模糊搜索
-		// 		// .find({ $or: [
-    //     //  	{'username': {$regex: ctx.query.queryName}}
-    //    	//	]
-    //     // }) 
-    //     .find(querys) // 模糊搜索
-		// 		.sort({ 'createTime': -1 }) // 排序，-1为倒序
-		// 		.skip(pageSize * (currentPage - 1)) // 跳过数
-		// 		.limit(pageSize) // 限制每页显示数
-		// 		.exec() // 执行sql语句
-		// 		.catch(err=>{
-		// 			ctx.throw(500, '服务器内部错误-getUser错误！')
-		// 		})
-		// 	total = await User
-		// 		// .count() // 5.2版本已废弃
-		// 		.countDocuments()
-		// 		.exec()
-		// 		.catch(err => {
-		// 			ctx.throw(500, '服务器内部错误-分页总数查询错误!')
-		// 		})
-		// }else{
-		// 	result = await User.find().exec().catch(err => {
-		// 		ctx.throw(500, '服务器内部错误-getUser错误！')
-		// 	})
-		// }
-		// if(result){
-		// 	handleSuccess({ ctx, msg: '数据获取成功！', 
-		// 		data: {
-		// 			pagination: {
-    //         total,
-    //         currentPage,
-    //         pageSize
-    //       },
-    //       list: result
-		// 		}
-		// 	})
-		// }else{
-		// 	handleError({ ctx, msg: '数据获取失败！' })
-		// }
 	}
 
 
@@ -228,74 +166,23 @@ class userController{
 			if (result) handleSuccess({ ctx, msg: '新增链接成功', data: result })
 			else handleError({ ctx, msg: '新增链接失败' })
 		}
-
-		// console.log(ctx.req.file)
-		// console.log(ctx.req.body)
-		// const { username, curPwd, surePwd, role, desc } = ctx.req.body
-		// const file = ctx.req.file
-		// if(!username){
-		// 	handleError({ ctx, msg: '用户名不能为空！' })
-		// 	if(file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// if (!curPwd){
-		// 	handleError({ ctx, msg: '密码不能为空！' })
-		// 	if(file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (curPwd != surePwd) {
-		// 	handleError({ ctx, msg: '两次密码输入不一致！' })
-		// 	if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// if (!role){
-		// 	handleError({ ctx, msg: '角色没有分配！' })
-		// 	if(file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (!desc) {
-		// 	handleError({ ctx, msg: '角色描述不能为空！' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// // 查询新增用户名是否已存在
-		// let oneUser = await User
-		// 	.findOne({ 'username': username })
-		// 	.exec() // 执行sql语句
-		// 	.catch(err => {
-		// 		ctx.throw(500, '服务器内部错误-findOneUser错误！')
-		// 	})
-		// if(oneUser === null){
-		// 	const user = new User({
-		// 		username,
-		// 		password: md5(curPwd),
-		// 		role: parseInt(role),
-		// 		desc: desc,
-		// 		avatar: file ? `upload/avatar/${file.filename}` : 'upload/avatar/default.png',
-		// 		createTime: new Date().getTime()
-		// 	})
-		// 	let result = await user.save().catch((err) => {
-		// 		ctx.throw(500, '服务器内部错误-addUser错误！')
-		// 	})
-		// 	handleSuccess({
-		// 		ctx, msg: '新增成功！',
-		// 		data: result
-		// 	})
-		// }else{
-		// 	if(file) fs.unlinkSync(file.path)
-		// 	handleError({ ctx, msg: '用户名已存在！' })
-		// }
 	}
 
 	// 编辑用户&修改图片
 	static async putUser(ctx){
 		const _id = ctx.params.id
-		const { username, role, desc, prePwd, newPwd, surePwd } = ctx.req.body
+		// const { username, role, desc, prePwd, newPwd, surePwd } = ctx.req.body
+		const { prePwd, newPwd, surePwd } = ctx.req.body
 		const file = ctx.req.file
-		console.log(ctx.params.id)
-		console.log(ctx.req.body)
-		console.log(ctx.req.file)
+		// console.log(ctx.params.id)
+		// console.log(ctx.req.body)
+		// console.log(ctx.req.file)
 
+		// 去除不更新的字段
+		delete ctx.req.body.prePwd
+		delete ctx.req.body.newPwd
+		delete ctx.req.body.surePwd
+		console.log(prePwd, newPwd, surePwd)
 		if (!_id) {
 			throw new CustomError(500, '无效参数')
 			return false
@@ -307,44 +194,42 @@ class userController{
 				throw new CustomError(500, '服务器内部错误')
 				return false
 			})
-		if (!prePwd) {
-			if (file) fs.unlinkSync(file.path)
-			throw new CustomError(500, '原始密码不能为空')
-			return false
+		if (prePwd || newPwd || surePwd){
+			if (!prePwd) {
+				if (file) fs.unlinkSync(file.path)
+				throw new CustomError(500, '原始密码不能为空')
+				return false
+			}
+			if (!newPwd) {
+				if (file) fs.unlinkSync(file.path)
+				throw new CustomError(500, '新密码不能为空')
+				return false
+			}
+			if (!surePwd) {
+				if (file) fs.unlinkSync(file.path)
+				throw new CustomError(500, '确认密码不能为空')
+				return false
+			}
+			if (oneUser.password != md5(prePwd)) {
+				if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
+				throw new CustomError(500, '原始密码输入有误！')
+				return false
+			}
+			if (newPwd == prePwd) {
+				if (file) fs.unlinkSync(file.path)
+				throw new CustomError(500, '新密码不能与原始密码相同')
+				return false
+			}
+			if (newPwd != surePwd) {
+				if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
+				throw new CustomError(500, '新密码两次输入不一致！')
+				return false
+			}
+			ctx.req.body.password = md5(newPwd)
 		}
-		if (!newPwd) {
-			if (file) fs.unlinkSync(file.path)
-			throw new CustomError(500, '新密码不能为空')
-			return false
-		}
-		if (!surePwd) {
-			if (file) fs.unlinkSync(file.path)
-			throw new CustomError(500, '确认密码不能为空')
-			return false
-		}
-		if (oneUser.password != md5(prePwd)){
-			if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-			throw new CustomError(500, '原始密码输入有误！')
-			return false
-		}
-		if (newPwd == prePwd) {
-			if (file) fs.unlinkSync(file.path)
-			throw new CustomError(500, '新密码不能与原始密码相同')
-			return false
-		}
-		if (newPwd != surePwd) {
-			if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-			throw new CustomError(500, '新密码两次输入不一致！')
-			return false
-		}
+		ctx.req.body.avatar = file ? `upload/avatar/${file.filename}` : oneUser.avatar
 		const result = await User
-			.findByIdAndUpdate(_id, { 
-				username,
-				password: md5(newPwd),
-				role,
-				desc,
-				avatar: file ? `upload/avatar/${file.filename}` : oneUser.avatar
-			}, { new: true }) // new: true ？？？
+			.findByIdAndUpdate(_id, ctx.req.body , { new: true }) // new: true ？？？
 			.catch(err => {
 				throw new CustomError(500, '服务器内部错误')
 				return false
@@ -372,102 +257,6 @@ class userController{
 			}
 			handleSuccess({ ctx, msg: '修改数据成功', data: result })
 		} else handleError({ ctx, msg: '修改数据失败' })
-
-		// const { id, username, role, desc, prePwd, newPwd, surePwd } = ctx.req.body
-		// const file = ctx.req.file
-		// let oneUser = await User
-		// 	.findOne({ '_id': id })
-		// 	.exec() // 执行sql语句
-		// 	.catch(err => {
-		// 		ctx.throw(500, '服务器内部错误-findOneUser错误！')
-		// 	})
-		// // console.log(oneUser)
-		// if (!username) {
-		// 	handleError({ ctx, msg: '用户名不能为空！' })
-		// 	if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// let aUser = await User
-		// 	.findOne({ 'username': username })
-		// 	.exec() // 执行sql语句
-		// 	.catch(err => {
-		// 		ctx.throw(500, '服务器内部错误-findOneUser错误！')
-		// 	})
-		// // console.log(aUser)
-		// if (aUser != null){
-		// 	handleError({ ctx, msg: '用户名已存在' })
-		// 	if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// if (!prePwd) {
-		// 	handleError({ ctx, msg: '原始密码不能为空' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (oneUser.password != md5(prePwd)){
-		// 	handleError({ ctx, msg: '原始密码输入有误！' })
-		// 	if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// if (!newPwd) {
-		// 	handleError({ ctx, msg: '新密码不能为空' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (newPwd == prePwd) {
-		// 	handleError({ ctx, msg: '新密码不能与原始密码相同' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (!surePwd) {
-		// 	handleError({ ctx, msg: '确认密码不能为空' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (newPwd != surePwd) {
-		// 	handleError({ ctx, msg: '新密码两次输入不一致！' })
-		// 	if (file) fs.unlinkSync(file.path) // 验证失败删除已上传的头像
-		// 	return false
-		// }
-		// if (!role) {
-		// 	handleError({ ctx, msg: '角色没有分配！' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// if (!desc) {
-		// 	handleError({ ctx, msg: '角色描述不能为空！' })
-		// 	if (file) fs.unlinkSync(file.path)
-		// 	return false
-		// }
-		// let result = await User.findByIdAndUpdate(id, {
-		// 	username,
-		// 	password: md5(newPwd),
-		// 	role,
-		// 	desc,
-		// 	avatar: file ? `upload/avatar/${file.filename}` : oneUser.avatar
-		// }, { new: true })
-		// 	.exec() // 执行查询，并将查询结果传入回调函数,可以传人一个函数，会返回成为一个完整的 promise 对象
-		// 	.catch((err) => {
-		// 		ctx.throw(500, '服务器内部错误-findByIdAndUpdate错误!');
-		// 	})
-		// // 有新图片上传 更新完毕后将老图删除
-		// // console.log(file.path)
-		// // 	要过滤掉默认图片，不然会被删除
-		// if (file && !oneUser.avatar.includes('default.png')) {
-		// 	// 先读取头像看是否存在,确保头像不存在的去删除的异常
-		// 	fs.readFile(`static/${oneUser.avatar}`, (err, data)=>{
-		// 		// 读取文件失败/错误
-		// 		if (err) {
-		// 			// throw err;
-		// 			console.log(err)
-		// 		} else {
-		// 			// 读取文件成功
-		// 			console.log(data)
-		// 			fs.unlinkSync(`static/${oneUser.avatar}`)
-		// 		}
-		// 	})
-		// }
-		// handleSuccess({ ctx, msg: '修改成功！', data: result })
 	}
 
 	// 删除用户
@@ -506,36 +295,6 @@ class userController{
 			}
 			handleSuccess({ ctx, msg: '删除成功', data: result })
 		} else handleError({ ctx, msg: '删除失败' })
-
-		// const _id = ctx.params.id
-		// if (!_id) {
-		// 	handleError({ ctx, msg: '参数无效' })
-		// 	return false
-		// }
-		// const result = await User
-		// 	.findByIdAndRemove(_id)
-		// 	.catch(err => {
-		// 		ctx.throw(500, '服务器内部错误-deleteUser错误！')
-		// 	})
-		// if (result) {
-		// 	// fs.unlinkSync(`./static/${result.avatar}`) // 删除头像
-		// 	if (result.avatar){
-		// 		// 先读取头像看是否存在
-		// 		fs.readFile(`static/${result.avatar}`, (err, data) => {
-		// 			// 读取文件失败/错误
-		// 			if (err) {
-		// 				// throw err;
-		// 				console.log(err)
-		// 			}else{
-		// 				// 读取文件成功
-		// 				console.log(data);
-		// 				fs.unlinkSync(`static/${result.avatar}`)
-		// 			}
-		// 		})
-		// 	} 
-		// 	handleSuccess({ ctx, msg: '删除成功！', data: result })
-		// }
-		// else handleError({ ctx, msg: '删除失败！' })
 	}
 
 	// 用户退出 （前端直接退，无需后台退出）
